ml/training/scripts/train_document_classifier.py
#!/usr/bin/env python3
"""
Training script for document classification model
Uses LayoutLMv3 or similar document understanding model
"""

import os
import yaml
import mlflow
import torch
from pathlib import Path
from transformers import (
    LayoutLMv3ForSequenceClassification,
    LayoutLMv3Processor,
    TrainingArguments,
    Trainer
)
from datasets import load_dataset
import structlog

logger = structlog.get_logger()

def load_config(config_path: str):
    """Load training configuration"""
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)

def setup_mlflow(config):
    """Initialize MLflow tracking"""
    mlflow.set_tracking_uri(config['logging']['mlflow_tracking_uri'])
    mlflow.set_experiment(config['logging']['experiment_name'])

def prepare_datasets(config):
    """Load and prepare training datasets"""
    logger.info("Loading datasets...")

    # TODO: Implement actual dataset loading
    # For now, return placeholder
    train_dataset = None
    val_dataset = None
    test_dataset = None

    return train_dataset, val_dataset, test_dataset

def create_model(config):
    """Initialize model and processor"""
    logger.info("Initializing model...",
                pretrained=config['model']['pretrained'])

    processor = LayoutLMv3Processor.from_pretrained(
        config['model']['pretrained']
    )

    model = LayoutLMv3ForSequenceClassification.from_pretrained(
        config['model']['pretrained'],
        num_labels=config['model']['num_labels']
    )

    return model, processor

def compute_metrics(eval_pred):
    """Compute evaluation metrics"""
    predictions, labels = eval_pred
    predictions = predictions.argmax(axis=-1)

    # Calculate accuracy, precision, recall, F1
    # TODO: Implement detailed metrics
    accuracy = (predictions == labels).mean()

    return {
        'accuracy': accuracy
    }

def train(config_path: str):
    """Main training function"""
    logger.info("Starting training", config=config_path)

    # Load configuration
    config = load_config(config_path)

    # Setup MLflow
    setup_mlflow(config)

    # Start MLflow run
    with mlflow.start_run():
        # Log parameters
        mlflow.log_params({
            'model_type': config['model']['type'],
            'learning_rate': config['training']['learning_rate'],
            'batch_size': config['training']['batch_size'],
            'epochs': config['training']['epochs']
        })

        # Prepare datasets
        train_dataset, val_dataset, test_dataset = prepare_datasets(config)

        # Create model
        model, processor = create_model(config)

        # Training arguments
        training_args = TrainingArguments(
            output_dir=config['output']['checkpoint_dir'],
            num_train_epochs=config['training']['epochs'],
            per_device_train_batch_size=config['training']['batch_size'],
            per_device_eval_batch_size=config['training']['batch_size'],
            learning_rate=config['training']['learning_rate'],
            weight_decay=config['training']['weight_decay'],
            warmup_steps=config['training']['warmup_steps'],
            logging_steps=config['logging']['log_interval'],
            evaluation_strategy="steps",
            eval_steps=config['logging']['eval_steps'],
            save_steps=config['logging']['save_steps'],
            load_best_model_at_end=True,
            fp16=config['training']['mixed_precision'],
            gradient_accumulation_steps=config['training']['gradient_accumulation_steps'],
            max_grad_norm=config['optimization']['max_grad_norm']
        )

        # Initialize trainer
        trainer = Trainer(
            model=model,
            args=training_args,
            train_dataset=train_dataset,
            eval_dataset=val_dataset,
            compute_metrics=compute_metrics
        )

        # Train
        logger.info("Training started...")
        train_result = trainer.train()

        # Log metrics
        mlflow.log_metrics(train_result.metrics)

        # Evaluate on test set
        logger.info("Evaluating on test set...")
        test_metrics = trainer.evaluate(test_dataset)
        mlflow.log_metrics({f"test_{k}": v for k, v in test_metrics.items()})

        # Save model
        logger.info("Saving model...", output_dir=config['output']['model_dir'])
        model.save_pretrained(config['output']['model_dir'])
        processor.save_pretrained(config['output']['model_dir'])

        # Log model to MLflow
        mlflow.transformers.log_model(
            transformers_model={
                "model": model,
                "processor": processor
            },
            artifact_path="model"
        )

        logger.info("Training completed successfully!")

if __name__ == "__main__":
    import argparse

    parser = argparse.ArgumentParser()
    parser.add_argument(
        "--config",
        type=str,
        default="configs/document_classification.yaml",
        help="Path to training config file"
    )

    args = parser.parse_args()
    train(args.config)
