ml/training/requirements.txt
# ML Training Pipeline
torch==2.2.2
torchvision==0.17.2
transformers==4.40.0
datasets==2.18.0
accelerate==0.29.3

# ML frameworks
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0

# Data processing
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0

# MLOps
mlflow==2.12.1
dvc==3.49.0
wandb==0.16.6

# Experiment tracking
optuna==3.6.1
tensorboard==2.16.2

# Data versioning
great-expectations==0.18.12

# Model evaluation
scikit-plot==0.3.7
seaborn==0.13.2
matplotlib==3.8.4

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1
tqdm==4.66.2
jsonlines==4.0.0

# Testing
pytest==8.1.1
pytest-cov==5.0.0
